feat(accounts): move accounts ViewModels into Accounts namespace.
