+++ app/Http/Controllers/AccountController.php
@@ -15,8 +15,8 @@ use App\Http\Requests\Accounts\IndexAccountRequest;
 use App\Http\Requests\Accounts\UpdateAccountRequest;
 use App\Models\Account;
 use App\Resources\Accounts\AccountResource;
-use App\ViewModels\GetAccountViewModel;
-use App\ViewModels\PaginateAccountsViewModel;
+use App\ViewModels\Accounts\GetAccountViewModel;
+use App\ViewModels\Accounts\PaginateAccountsViewModel;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Facades\Gate;
