+++ app/ViewModels/Accounts/GetAccountViewModel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\ViewModels;
+namespace App\ViewModels\Accounts;
 
 use App\Models\Account;
 use App\Resources\Accounts\AccountResource;
